Graphic representation of cards adjusted
Have returned to defining the path to background image in SCSS file,
and am manually removing the surplus instances of 'build/' in the
compiled CSS file.

Defining the background image via inline CSS in CardGraphic class
led to faulty paths when attempting to access the image from sub
routes.

Removing the instances of 'build/' in the compiled CSS file has also
made possible the usage of Fontawesome characters in the footer, so
those have been added for social links.