Added route for Draw and associated functionality
Added route card/deck/draw for drawing a card and displaying it along with number
of cards remaining in deck.

Added route card/deck/create-and-shuffle/{source} for creating and
shuffling a new deck of cards, with the possibility of choosing
redirection route depending on the source route.

Added functionality to the route card/deck/shuffle to upon access
create a new deck of cards when the number of cards in the session
is 0.